pgpool-II could not detach multiple node when message kinds were mismatched.
For example:
  node-1(master): success
  node-2: fail
  node-3: success
  node-4: fail
  node-5: success

node-2 was only detached.

Now it determine detached nodes by "majority voting rule".
And it can detach them at the same time.

For example:
  node-1(master): success
  node-2: fail
  node-3: success
  node-4: fail
  node-5: success

node-2 and node-4 are detached because they are minority.

  node-1(master): success
  node-2: fail
  node-3: fail
  node-4: success

The number of the group which gain "success" is same as the number of
the group which gain "fail". In this case, node-2 and node-3 are
detached because these don't belong to the group which master belongs
to.