Prpare for receiving parameter status packets.
If postgresql.conf was changed and reloaded, PostgreSQL sends
parameter status packets asynchronously even those changed parameters
are not one of session_authorization etc., which is marked as
GUC_REPORT in guc.c(GUC_REPORT means parameter status packet is sent
if its value is changed).

After receiving those packets, they are discarded. I think this is
harmless since most of changes are not affected to the frontend
behavior.