Fix ProcessFrontendResponse() handling of Parse message in master/slave mode. If it's inside transaction, enable master_slave_dml, which means execute parse on master only. This will prevent unwanted parase executed on slave, which holds at least AccessShareLock on tables until transaction commits. It seems unamed statement does not release the lock until execute.
Please note that, even with this fix, bind state will remain on slave
if JDBC issues parse with autocommit off. This is because BEGIN
parsed, bind on both master and slave but execute only on
master. However, this is not harmfull since parse/bind does not hold
any locks on tables.

Also add some logging statemets.