Fix bug with debug_level directive. Patch contributed by Gilles Darold. Followings are his explanation.
I found a bug when playing with pgpool yesterday. The problem is that
enabling debug_level to 1 into pgpool.conf doesn't has any effect as it
is overriden in the main.c file after getting configuration so that it
is only enabled when the -d command line option is present.