Fix bug which cannot use the cursors of JDBC driver on standby node.
In master/slave mode, the transaction commands come to be sent to
all nodes by this fix because the cursors of JDBC driver cannot use
without transactions.

But if the transaction command is "BEGIN READ WRITE" or "START TRANSACTION
READ WRITE", pgpool-II sends "BEGIN" instead of it to standby node.
The reason is because standby nodes disallow "READ WRITE" option.