Fix pool_process_query() rus into infinite loop reported in this thread: http://www.pgpool.net/pipermail/pgpool-general/2011-December/000099.html
In the code path the case when receive buffer of primary is empty and
one of standbys sends spontaneously packet to pgpool (thus their
recieve buffers have data) is not considered. This could happen when,
for example, reloading postgresql.conf. The fix is, read one packet
from standby's receive buffer and discard it.

Note that the problem started with pgpool-II 3.0.4 (released on
2011/6/1).