Fix to use memory cache of extended queries in a transaction, too.
When a transaction is comitted, query_context->temp_cache->query is used
to create md5 hash to search for query cache.
So overwrite the query text in temp cache to the one with the hex of bind message.
If not, md5 hash will be created by the query text without bind message, and
it will happen to find cache never or to get a wrong result.