Fix bug with query cache returning incorrect data in some cases when a persistent table and temp table have same name.
Here is a sequence to trigger the bug:
1) CREATE TABLE t1(i int);	-- create a persistent table
2) INSERT INTO t1 VALUES(1);
3) SELECT * FROM t1;		-- query cache entry created
4) CREATE TEMP TABLE t1(i int);	-- create a temp table
5) SELECT * FROM t1;		-- query cache entry mistakenly created!

Problem is #3 creates relcache entry for t1, and #5 incorrecly uses it
and believes that temp table t1 is not a temp table. To solve the
problem, new function discard_temp_table_relcache() added and it gets
called in ReadyforQuery if CREATE TEMP TABLE executed to discard
relcache for temp table info.