Fix long standing bug "portal not found" error when replication delay is too much in streaming replication mode.  The bug had been there since the delay threshold was introduced.
We changed destination DB node if delay threshold exceeds in bind,
describe and execute. However, if parse sends to different node, bind,
describe or execute will fail because no parsed statement or portal
exists.  Solution is, not to send to different parse node even
if delay threshold is too much.