Avoid a possible hang during the active pgpool exits.
When exiting, the active pgpool brings down the virtual IP and then
sends a packet to other pgpools. However, the packet sometimes is sent
before the virtual IP is brought down completely. In this case the packet
sender is set to this IP. When the IP is brought down before other pgpools
respond, the active pgpool can not recieve the response, and hang up.

In this fix, the active pgpool confirms that the virtual IP is brought
down before sending the packet.