Fix bug with on memory query cache in handling UPDATE/DELETE with table alias.
If UPDATE/DELETE is with table alias (UPDATE t1 AS foo...) pgpool
thinks the table name is "t1 AS foo" and fails to invalidate query
cache. This is caused by _outRangeVar() called from nodeToString()
which generates a query string from RangeVar node in raw parse
tree. The solution is removing "AS foo" part from the output of the
string. But reported in bugtrack #56.