Fix memqcache segfaults reported in bug#63.
When a query executes and the query result is too large,
query_context->temp_cache is discarded the query result is not
committed to cache storage when the query completed. In extended query
mode, it is possible that the same query context is used for
subsequent bind/execute cycle through reusing the portal. Problem is,
query_context->temp_cache is already gone and a segfault is caused by
accessing the pointer. To fix this, when query_context->temp_cache is
discarded, set the pointer to NULL and whenever access
query_context->temp_cache, check it is NULL or not.