Enhance pgpool_setup and regression test suite.
Patch contributed by Muhammad Usama. Below is the comment from him:

I tried to make some changes in the regression script, mainly because the
script uses lot of hard coded paths and regression starts reporting
failures if the pgpool-II or PostgreSQL are not found at the default path.

Please find the attached patch which enables the regress.sh script to
install the pgpool on fly in temp directory to run the tests on it
(optional), and now the regress script takes some additional arguments to
specify the paths of required components.

The list of all arguments accepted by regress.sh script as implemented by
the attached patch are
pgpool2/test/regression$ sh regress.sh -?
Usage:
  regress.sh: [Options]... [test_name]

Options:
  -p   DIRECTORY           Postgres installed directory
  -b   DIRECTORY           pgbench installed directory, if different from
Postgres installed directory
  -i   DIRECTORY           pgpool installed directory, if already installed
pgpool is to be used for tests
  -m   install/noinstall   make install pgpool to temp directory for
executing regression tests [Default: install]
  -j    FILE               Postgres jdbc jar file path
  -?                       print this help and then exit