Fix pgbench path detection.
-b option specifies pgbench installed directory. But actually
regress.sh main script sets pgbench full path. I think what the main
script does is appropreate because it gives more flxibility. If -b is
not specifiedm, first search pgbench under PostgreSQL installation
directory. If not found, try to locate in command search path. If not
found, the gives up.

Also fix the behavior when PostgreSQL installation directory is not
found. It just exited in a sub shell and it continued to run.