Fix possible pgpool-recovery crash when used with pre-3.4 environment.
Between 3.3 and 3.4 the number of parameters are changed from 3 to
4. Before pgpool-recovery checks the number of arguments by using
PointerGetDatum(PG_GETARG_TEXT_P(3)) returns null or not.  This is
wrong and now we use PG_NARGS() macro.

Also pgpool-recovery's Makefile now installs pgpool_recovery--1.1.sql,
rather than pgpool_recovery--1.0.sql to deal with the changes of
number of arguments.