Fix an infinite loop of error at new active pgpool when virtual IP switching
When active pgpool stops and standby pgpool becomes new active, this brings
virtual IP up and confirm this by ping command. In previous, pgpool handle
exit code of ping wrongly and goes into an infinite loop.

In addition, the old active pgpool also hangs since this waits response
from the new active pgpool, which is in the infinite loop and not able
to respond.

This problem is reported by Qian Peng in [pgpool-hackers: 520].