Fix regression test 057 hung if run against PostgreSQL 9.4 or later.
PostgreSQL 9.4 or later changed libpq behavior: NULL username for
connection string does not cause an error (which is expected in the
test). As a result psql prompts and waits for user's input which
causes regression test hang if executed against PostgreSQL 9.4 or
later. To fix this, "\q" is implicitly provided to psql session.

Now the test (pgpool segafult) is only performed against pre 9.4. 9.4
or later happily passes the test.

Regarding libpq change see:
http://www.postgresql.org/message-id/20140708.111307.408695214582895632.t-ishii@sraoss.co.jp
for more details.