Add new directive "connect_timeout".
This is for timeout value in milliseconds before giving up to connect
to backend using connect() system call in
connect_inet_domain_socket_by_port() which is used for health check
and creating backend connection.  Since we use non-blocking connect(),
we use select() if connect() returns "in progress" state. Before the
timeout value for select() was fixed 1 second, which is not long
enough for flaky network such as AWS.  So new knob is added to control
the timeout value.