Back porting coverity issues fix patch from master branch. Fix coverity issue:1111377 Unchecked return value Fix coverity issue:1111381 Logically dead code Fix coverity issue:1222992 Unchecked return value Fix coverity issue:1222993 Unchecked return value Fix coverity issue:1222996 Printf arg type mismatch Fix coverity issue:1222997 Printf arg type mismatch
For the coverity issue 1222996 and 1222997 the return type of NameListToString function
is changed from string* to char*, which is the signature of this function in PostgreSQL source.
and appropriate changes have been made in the code to incorporate this function return type change.

This checkin also fixes the memory leak in NameListToString Function and adds the prototype of same in pool_string.h, That
was causing the compiler warning.