Back porting coverity issues fix patch from master branch. Fix coverity issue:1111377 Unchecked return value Fix coverity issue:1111381 Logically dead code Fix coverity issue:1222992 Unchecked return value Fix coverity issue:1222993 Unchecked return value Fix coverity issue:1222996 Printf arg type mismatch Fix coverity issue:1222997 Printf arg type mismatch
For the coverity issue 1222996 and 1222997 the return type of NameListToString function
is changed from string* to char*, and defination of the which is the return type of the function in PostgreSQL source.
and appropriate changes have been made in the code to incorporate this function return type change.

For coverity issue:1222992 1222993 the return type of do_query() function is changed to void from POOL_STATUS.
The function’s return value became insignificant after the introduction of elog API,
After the API do_query() only returns in a successful scenarios and throws an ereport
in case of any error within the function.