First cut of the implementation of SELECT count stats.
The SELECT counter is added to "show pool_nodes". Here is an example:
show pool_nodes;
 node_id | hostname | port  | status | lb_weight |  role   | select_cnt
---------+----------+-------+--------+-----------+---------+------------
 0       | /tmp     | 11002 | 2      | 0.500000  | primary | 9231
 1       | /tmp     | 11003 | 2      | 0.500000  | standby | 9469
(2 rows)

For now, the counter is in shared memory are and resets upon starting
up and there's no functionality without start/stop of
pgpool-II. Probably new pcp commands is necessary to reset the counter
without restarting.