Second part of overhauling the PCP library and utilities.
This commit contains the following changes

-- pgpool-II can now handle multiple simultaneous PCP commands.

-- pcp_child process now only listens for new connections and manages pcp_worker
children. Responsibility of authenticating PCP clients and processing of the
PCP client request is delegated to new PCP worker type process.
Which are spawned on demand when the new command request is received

-- This commit changes the way how smart shutdown request is handled by the
PCP process. Previously pcp_child was not caring about if smart or fast shutdown
request was issued, and kills the process instantly. This commit ensures that if
a smart shutdown request was received than the pcp_child stops receiving the
new PCP connections, but the current command in execution executes at full
before the processing of the shutdown request.

-- Restart pcp_child request now waits for the current command to complete
before restarting the process, instead of immediately killing the PCP process.

-- Removes all occurrences and handling of unused pcp_timeout from code and
command line arguments of PCP utilities. (Documentation changes are still a TODO)

-- Removes the global variable run_as_pcp_child which was already never used.

-- pgpool-II can now execute multiple concurrent PCP commands except for
pcp_recovery_node command.(only one pcp_recovery_node is allowed at a time).

-- Since pcp_conf_file is the global parameter so instead of passing it to the
pcp_child main process as an argument it is extended in the pcp_worker file.

-- Removes now unwanted utils/pcp/pcp_error.c and  utils/pcp/pcp_timeout.c files.

-- Commit also contains some miscellaneous code cleanups.