Next in the series of overhauling the PCP library and utilities.
This list of some more prominent changes is as follows.

-- As currently pgpool-II has 12 PCP client utilities and almost 90% of the code
for these utilities was common, So this commit creates a single source file for
all the utilities, to make managing and maintenance of the utilities little easier.
New source file "src/tools/pcp/pcp_frontend_client.c" added by this commit is
now used for creating all PCP utilities.

-- Syntax is changed for all PCP utilities and PCP utilities now accepts long-options.
Also, almost all arguments to the utilities are assigned a default value which
is used if the particular argument is not supplied by the user.

-- The commit also solves the security issue existed in all PCP utilities,
which was due to the fact that all PCP commands accepts the password as an
argument which has a potential of being captured by 'ps' command.
Now PCP applications do not take password in an argument and it can be either
supplied by password prompt or utilities can read the password
from ". pcppass" file, which should be located in User's home directory
or can be specified by  PCPPASSFILE environment variable,
The password file format is "hostname:port:username:password"