Allow to open pool_passwd file in read only mode by pgpool-II main process.
Since pgpool-II process does not write to the file, it should be
allowed to open the file in read only mode by the pgpool-II
process. To implement this, now pool_init_pool_passwd has an
additional parameter "mode" which specifies in which mode caller want
to open pool_passwd. Also new static variable "pool_passwd_mode" is
introduced to remember in which mode the current file descriptor has
been created.