Fix for pgpool-II emits a fatal error message when clear text password auth is used.
The problem was introduced in pgpool-II after inclusion of exception manager.
EOF on the frontend connection during client authentication should not be treated
as an error when the password authentication method is used. psql disconnects
from the server in such case to get the password from the terminal,
reconnects and sends the read password to server.

The check-in also enhances the error messages for authentication failures
by concatenating backend errors to pgpool error report.